Remove test of control API in favor of poll-api
